+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - [**Improved**] **Technical**: Versioning now only handled by pyproject
 - [**Changed**] **Technical**: Adapted discord.py's [asyncio changes](https://gist.github.com/Rapptz/6706e1c8f23ac27c98cee4dd985c8120)
 - [**Improved**] You can now set Welcome and Farewell message using Modal
+- [**Changed**] **Technical**: Splitting meta into several subcogs
 
 # v3.4