+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # v3.4
 
+## 3.4.7
+- [**Fixed**] Fixed NSFW commands (`nekos.life` NSFW endpoints is dead)
+
 ## 3.4.6
 - [**Fixed**] Getting 0 no longer possible on normal dice roll